Add support for building cesium-native with Unity Web

+++ Cesium3DTilesContent/src/BatchTableToGltfStructuralMetadata.cpp
@@ -1014,7 +1014,10 @@ void copyStringsToBuffers(
     for (const auto& str : arrayMember.GetArray()) {
       OffsetType byteLength = static_cast<OffsetType>(
           str.GetStringLength() * sizeof(rapidjson::Value::Ch));
-      std::memcpy(valueBuffer.data() + offset, str.GetString(), byteLength);
+      std::memcpy(
+          valueBuffer.data() + offset,
+          str.GetString(),
+          size_t(byteLength));
       std::memcpy(
           offsetBuffer.data() + offsetIndex * sizeof(OffsetType),
           &offset,
@@ -1075,8 +1078,7 @@ void updateStringArrayProperty(
     ++it;
   }
 
-  const uint64_t totalByteLength =
-      totalCharCount * sizeof(rapidjson::Value::Ch);
+  const size_t totalByteLength = totalCharCount * sizeof(rapidjson::Value::Ch);
   std::vector<std::byte> valueBuffer;
   std::vector<std::byte> stringOffsetBuffer;
   PropertyComponentType stringOffsetType = PropertyComponentType::None;
